Revert "add opt cmd swap option based on d20ba5d17969a109b5acc73c71a31237de9985ec and subsequent changes"
This doesn't work yet because there's no way to actually set a value for this pref -- it isn't added to the guichan GUI
and it overwrites the whole config file.
And while it could be implemented, actually the common SDL launch screen configuration already has key remapping that can
be used for this so it's basically unnecessary to have a dedicated feature.

This reverts commit d47ff4b3ec33dd7f0aec0da93ff05e985cc1cb37.